requirements.txt
boto3
botocore
ipykernel
ipython
jupyter
jupyterlab==3.5.2
numpy
pandasWW
pyarrow
pyspark==3.4.0
requests
kafka-python
